+++ cal3d/src/cal3d/vector.cpp
@@ -377,7 +377,7 @@ void CalVector::clear()
 
 float CalVector::length()
 {
-  return sqrt(x * x + y * y + z * z);
+  return (float)sqrt(x * x + y * y + z * z);
 }
 
  /*****************************************************************************/
@@ -392,7 +392,7 @@ float CalVector::normalize()
 {
   // calculate the length of the vector
   float length;
-  length =sqrt(x * x + y * y + z * z);
+  length = (float) sqrt(x * x + y * y + z * z);
 
   // normalize the vector
   x /= length;
